wai-aria/role/region-roles.html

<!doctype html>
<html>
<head>
  <title>Region Role Verification Tests</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="/resources/testdriver.js"></script>
  <script src="/resources/testdriver-vendor.js"></script>
  <script src="/resources/testdriver-actions.js"></script>
</head>
<body>

<div role="region" id="no_label">x</div>
<div role="region" id="label" aria-label="x">x</div>

<script>

promise_test(async t => {
  const role = await test_driver.get_computed_role(document.getElementById('no_label'));
  assert_equals(role, "generic");
}, "region with no label");

promise_test(async t => {
  const role = await test_driver.get_computed_role(document.getElementById('label'));
  assert_equals(role, "region");
}, "region with label");

</script>
</body>
</html>
